+++ CHANGELOG.md
@@ -2,4 +2,4 @@
 
 ## 1.0.0
 
-- Initial release of the ADS-B rtl1090 sensor.
+- Initial release of the ADS-B tar1090 sensor.